pyproject.toml
[project]
name = "model-context-protocol"
version = "0.1.0"
description = ""
authors = [
    {name = "Arka Patra",email = "arkapatra2009@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "mcp[cli] (>=1.6.0,<2.0.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "langchain (>=0.3.22,<0.4.0)",
    "langchain-core (>=0.3.49,<0.4.0)",
    "langchain-openai (>=0.3.11,<0.4.0)",
    "langchain-aws (>=0.2.18,<0.3.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "pymongo (>=4.11.3,<5.0.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "websocket (>=0.2.1,<0.3.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
